[TASK] Bootstrap for Unit Tests
Until now we still needed the phpunit extension
to be installed and cli access to be configured
to execute unit tests.

This change introduces a bootstrap file which
can be used to execute the unit test suite
without the extension.

It is of course still possible to use the PHPUnit
and VHS framework files present in the extension.

This change also splits user initialization into
separate bootstrap methods and allows the PATH_site
and PATH_thisScript constants to be set outside
the bootstrap.

Activate Travis CI to make use of this.

Change-Id: I43f51896fd9895ead8d0084f37d34f3347f1c5e5
Resolves: #51436
Releases: 6.2